EPMEDU-4493: Refactor GuestAlertWrapperView and update localization
- Moved GuestAlertWrapperView to a new file in Common/UI and refactored it to use localized strings.
- Updated Localizable.strings and Strings.swift to add new keys for 'register' and 'registerOrLogin'. - Removed hardcoded strings from MainCoordinator and CustomAlertView, centralizing guest alert logic and improving maintainability.
