+++ UIComponents/Sources/UIComponents/SwiftUIComponents/CustomAlertView.swift
@@ -155,44 +155,6 @@ extension CustomAlertView {
     }
 }
 
-// MARK: - GuestAlertWrapperView
-public struct GuestAlertWrapperView: View {
-    @State private var showAlert = false
-    let dismiss: () -> Void
-
-    public init(dismiss: @escaping () -> Void) {
-        self.dismiss = dismiss
-    }
-
-    public var body: some View {
-        CustomAlertView(
-            viewModel: CustomAlertView.ViewModel(
-                title: "Your are logged in as a guest. Register quickly with basic details for a better experience.",
-                message: nil,
-                primaryButtonTitle: "Register",
-                secondaryButtonTitle: "Cancel"
-            ) { _ in
-                dismissAlert()
-            },
-            isPresented: showAlert
-        )
-        .onAppear {
-            DispatchQueue.main.async {
-                showAlert = true
-            }
-        }
-    }
-
-    private func dismissAlert() {
-        withAnimation {
-            showAlert = false
-        }
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-            dismiss()
-        }
-    }
-}
-
 // MARK: - Preview
 struct CustomAlert_Previews: PreviewProvider {
     static let designEngine: StyleEngine = StyleDefaultEngine()
