+++ animeal/res/sheet.csv
@@ -110,4 +110,4 @@ Please contact us at %@","ვერ იპოვეთ თქვენი შე
 "verification.resendCode.titleTime","","Resend code in","კოდის ხელახლა გაგზავნა","","","","","","","","","","","","","","","","","","","","","","",""
 "loginScreen.continueAsGuest","","Continue as a Guest","შესვლა სტუმრად","","","","","","","","","","","","","","","","","","","","","","",""
 "loginScreen.registerOrLogin","","Register or log in to unlock advanced features!","","","","","","","","","","","","","","","","","","","","","","","",""
-"loginScreen.register","","Register","","","","","","","","","","","","","","","","","","","","","","","",""
+"action.register","","Register","","","","","","","","","","","","","","","","","","","","","","","",""
